IGNITE-20447 Fix after review by Slava Koptilin.
