+++ modules/failure-handler/build.gradle
@@ -22,8 +22,12 @@ apply from: "$rootDir/buildscripts/java-integration-test.gradle"
 apply from: "$rootDir/buildscripts/java-test-fixtures.gradle"
 
 dependencies {
-    implementation project(':ignite-api')
     implementation project(':ignite-core')
+
+    testImplementation libs.mockito.core
+    testImplementation libs.mockito.junit
+
+    testImplementation(testFixtures(project(':ignite-core')))
 }
 
 description = 'ignite-failure-handler'
